src/main/java/com/myretail/model/ProductPrice.java
package com.myretail.model;

import java.math.BigDecimal;

import javax.validation.constraints.Min;

import org.hibernate.validator.constraints.NotEmpty;
import org.springframework.data.cassandra.mapping.Column;
import org.springframework.data.cassandra.mapping.PrimaryKey;
import org.springframework.data.cassandra.mapping.Table;

@Table( "dylanmyretail.products" )
public class ProductPrice {

	@PrimaryKey
	private Long productId;
	
	@NotEmpty
	@Min(value = 0)
	@Column("price")
	private BigDecimal price;
	
	@Column("currency_code")
	private String currencyCode;

	public Long getProductId() {
		return productId;
	}

	public void setProductId(Long productId) {
		this.productId = productId;
	}

	public BigDecimal getPrice() {
		return price;
	}

	public void setPrice(BigDecimal price) {
		this.price = price;
	}

	public String getCurrencyCode() {
		return currencyCode;
	}

	public void setCurrencyCode(String currencyCode) {
		this.currencyCode = currencyCode;
	}
	
}
